pyproject.toml
[project]
name = "hovorunv2"
version = "0.1.0"
description = "The ultimate wingman for your Telegram group chats — making them smarter, faster, and way more fun."
readme = "README.md"
authors = [
    { name = "TwilightSparkle42", email = "jaden.corr42@gmail.com" }
]
license = { text = "BSD-3-Clause" }
keywords = ["telegram", "bot", "aiogram", "chat-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications :: Chat",
]
requires-python = ">=3.14"
dependencies = [
    "aiogram",
    "pydantic",
    "pydantic-settings",
    "aiohttp",
    "redis",
    "sqlalchemy",
    "aiosqlite",
    "greenlet",
    "alembic",
    "emoji-country-flag",
    "yt-dlp",
    "instaloader",
    "playwright",
    "beautifulsoup4",
    "dishka",
]

[project.urls]
Repository = "https://github.com/HovorunBot/HovorunV2"
Issues = "https://github.com/HovorunBot/HovorunV2/issues"

[project.scripts]
hovorunv2 = "hovorunv2.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hovorunv2"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-alembic",
    "pytest-asyncio",
    "ruff",
    "ty",
]

[tool.ruff]
line-length = 120
target-version = "py314"
exclude = ["alembic"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D107",
    "D203",
    "D213",
    "COM812",
    "ISC001",
    "S101", # Use of assertions allowed by user intent not for business logic
    "PLR0913",
    "D104",
    "BLE001",
    "TC"
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["FBT"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"



[tool.ty.src]
exclude = ["alembic/**"]
